smart-infra-gis/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0

# GIS
GDAL==3.7.3
Fiona==1.9.5
Shapely==2.0.2
pyproj==3.6.1
geopandas==0.14.1

# AWS
boto3==1.29.7

# Monitoring
sentry-sdk==1.38.0